feat: Enable next, previous and random buttons
*feat: Enable next, previous and random buttons

I have managed to get the next, previous and random buttons working.
Previously, the issue was that chrome was preventing the browser from
running the onclick code. So instead, I initialised the button
functionality in the js with `getElementById` instead. So that works
now.

Need to work on favourites next. After that it should be ready for
release.

Closes #4.